requirements.txt
scikit-learn==1.1.3
pandas==1.5.3
seaborn==0.12.1
matplotlib==3.6.2
numpy==1.23.5
fastapi==0.88.0
uvicorn==0.20.0
python-dotenv==0.21.0
joblib==1.2.0
